imp de permisos de usuario, retoque addSocio y msg errror, retoque modal confirmar pago

+++ src/main/java/ar/edu/unnoba/proyecto_river_plate_junin/controller/IndexController.java
@@ -19,7 +19,7 @@ public class IndexController {
 
     @GetMapping("/")
     public String index(Model model){
-        int sociosTotales = socioService.contarSocios();
+        Long sociosTotales = socioService.contarSocios();
         int usuariosTotales = userService.contarUsers();
         int sociosActivos = socioService.contarSociosActivos();
         model.addAttribute("usuarios",usuariosTotales );
